refactor: move temporal helpers into visualization package
