+++ qfit_dockwidget.py
@@ -44,7 +44,7 @@ from .visualization.application import BackgroundConfig, LayerRefs
 from .atlas.export_task import BUILTIN_ATLAS_MAP_TARGET_ASPECT_RATIO
 from .providers.domain.provider import ProviderError
 from .providers.infrastructure.strava_provider import StravaProvider
-from .temporal_config import DEFAULT_TEMPORAL_MODE_LABEL, temporal_mode_labels
+from .visualization.application import DEFAULT_TEMPORAL_MODE_LABEL, temporal_mode_labels
 from .ui.dockwidget_dependencies import DockWidgetDependencies, build_dockwidget_dependencies
 from .ui.workflow_section_coordinator import WorkflowSectionCoordinator
 from .ui_settings_binding import UIFieldBinding, load_bindings, save_bindings
