+++ visualization/application/__init__.py
@@ -6,6 +6,15 @@ from .background_map_controller import (
     LoadBackgroundResult,
 )
 from .layer_gateway import LayerGateway
+from .temporal_config import (
+    DEFAULT_TEMPORAL_MODE_LABEL,
+    TEMPORAL_MODE_LABELS,
+    TemporalLayerPlan,
+    build_temporal_plan,
+    describe_temporal_configuration,
+    is_temporal_mode_enabled,
+    temporal_mode_labels,
+)
 from .visual_apply import (
     ApplyVisualizationRequest,
     BackgroundConfig,
@@ -19,11 +28,18 @@ __all__ = [
     "ApplyVisualizationRequest",
     "BackgroundConfig",
     "BackgroundMapController",
+    "DEFAULT_TEMPORAL_MODE_LABEL",
     "LayerGateway",
     "LayerRefs",
     "LoadBackgroundRequest",
     "LoadBackgroundResult",
+    "TEMPORAL_MODE_LABELS",
+    "TemporalLayerPlan",
     "VisualApplyRequest",
     "VisualApplyResult",
     "VisualApplyService",
+    "build_temporal_plan",
+    "describe_temporal_configuration",
+    "is_temporal_mode_enabled",
+    "temporal_mode_labels",
 ]
